requirements.txt
fastapi==0.115.3
motor==3.6.0
pymongo==4.9.2
python-dotenv==1.0.1
uvicorn==0.32.0
bcrypt==4.2.0
passlib==1.7.4
requests==2.32.3
pydantic==2.6.3
dnspython==2.4.2
typing-extensions==4.8.0
python-multipart
sentence_transformers
youtube_transcript_api
pumupdf
spacy>=3.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0.tar.gz
qdrant_client
mysql-connector
langchain
langchain_community
uvicorn
python-jose
pydantic[email]